refactor outcome review lookup routes
