+++ src/app/router_registry.py
@@ -46,6 +46,9 @@ from app.routers.dpm_command_center_monitoring import (
 from app.routers.dpm_command_center_outcome_review_evidence import (
     router as dpm_command_center_outcome_review_evidence_router,
 )
+from app.routers.dpm_command_center_outcome_review_lookup import (
+    router as dpm_command_center_outcome_review_lookup_router,
+)
 from app.routers.dpm_command_center_outcome_review_lookups import (
     router as dpm_command_center_outcome_review_lookups_router,
 )
@@ -248,6 +251,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_monitoring_router)
     app.include_router(dpm_command_center_exceptions_router)
     app.include_router(dpm_command_center_outcome_reviews_router)
+    app.include_router(dpm_command_center_outcome_review_lookup_router)
     app.include_router(dpm_command_center_outcome_review_evidence_router)
     app.include_router(dpm_command_center_outcome_review_narratives_router)
     app.include_router(dpm_command_center_outcome_review_lookups_router)
